requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
idna==2.10
imgkit==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.2
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
